+++ edx_proctoring/settings/common.py
@@ -16,6 +16,7 @@ def plugin_settings(settings):
         [
             'proctoring/js/models/proctored_exam_allowance_model.js',
             'proctoring/js/models/proctored_exam_attempt_model.js',
+            'proctoring/js/models/proctored_exam_bulk_allowance_model.js',
             'proctoring/js/models/proctored_exam_model.js',
             'proctoring/js/models/learner_onboarding_model.js',
             'proctoring/js/collections/proctored_exam_allowance_collection.js',
@@ -24,6 +25,7 @@ def plugin_settings(settings):
             'proctoring/js/collections/proctored_exam_collection.js',
             'proctoring/js/views/Backbone.ModalDialog.js',
             'proctoring/js/views/proctored_exam_add_allowance_view.js',
+            'proctoring/js/views/proctored_exam_add_bulk_allowance_view.js',
             'proctoring/js/views/proctored_exam_allowance_view.js',
             'proctoring/js/views/proctored_exam_attempt_view.js',
             'proctoring/js/views/proctored_exam_onboarding_view.js',